Update La-Vision processing

+++ scripts/la-vision/predict_blocks.py
@@ -3,31 +3,50 @@ import os
 from glob import glob
 
 import imageio.v3 as imageio
+import numpy as np
+
+from skimage.segmentation import watershed
+from skimage.measure import label
 from torch_em.util import load_model
 from torch_em.util.prediction import predict_with_halo
 
-INPUT_ROOT = "/mnt/vast-nhr/home/pape41/u12086/Work/my_projects/flamingo-tools/scripts/more-annotations/LA_VISION_M04"  # noqa
+
+def _get_files(sgn=True):
+    input_root = "/mnt/vast-nhr/home/pape41/u12086/Work/my_projects/flamingo-tools/scripts/more-annotations/LA_VISION_M04"  # noqa
+    input_root2 = "/mnt/vast-nhr/home/pape41/u12086/Work/my_projects/flamingo-tools/scripts/more-annotations/LA_VISION_M04_2"  # noqa
+    input_root3 = "/mnt/vast-nhr/home/pape41/u12086/Work/my_projects/flamingo-tools/scripts/more-annotations/LA_VISION_Mar05"  # noqa
+
+    input_root_ihc = "/mnt/vast-nhr/home/pape41/u12086/Work/my_projects/flamingo-tools/scripts/more-annotations/LA_VISION_Mar05-ihc"  # noqa
+
+    if sgn:
+        input_files = glob(os.path.join(input_root, "*.tif")) +\
+            glob(os.path.join(input_root2, "*.tif")) +\
+            glob(os.path.join(input_root3, "*.tif"))
+    else:
+        input_files = glob(os.path.join(input_root_ihc, "*.tif"))
+
+    return input_files
 
 
-def predict_blocks(model_path, name):
+def predict_blocks(model_path, name, sgn=True):
     output_folder = os.path.join("./predictions", name)
     os.makedirs(output_folder, exist_ok=True)
 
-    input_blocks = glob(os.path.join(INPUT_ROOT, "*.tif"))
+    input_blocks = _get_files(sgn)
 
-    model = load_model(model_path)
+    model = None
     for path in input_blocks:
+        out_path = os.path.join(output_folder, os.path.basename(path))
+        if os.path.exists(out_path):
+            continue
+        if model is None:
+            model = load_model(model_path)
         data = imageio.imread(path)
         pred = predict_with_halo(data, model, gpu_ids=[0], block_shape=[64, 128, 128], halo=[8, 32, 32])
-        out_path = os.path.join(output_folder, os.path.basename(path))
         imageio.imwrite(out_path, pred, compression="zlib")
 
 
 def _segment_impl(pred, dist_threshold=0.5):
-    import numpy as np
-    from skimage.segmentation import watershed
-    from skimage.measure import label
-
     fg, center_dist, boundary_dist = pred
     mask = fg > 0.5
 
@@ -37,22 +56,26 @@ def _segment_impl(pred, dist_threshold=0.5):
     return seg
 
 
-def check_segmentation(name):
+def check_segmentation(name, sgn):
     import napari
 
-    input_blocks = sorted(glob(os.path.join(INPUT_ROOT, "*.tif")))
+    input_files = _get_files(sgn)
 
     output_folder = os.path.join("./predictions", name)
-    pred = sorted(glob(os.path.join(output_folder, "*.tif")))
 
-    for path, pred_path in zip(input_blocks, pred):
+    for path in input_files:
         image = imageio.imread(path)
+        pred_path = os.path.join(output_folder, os.path.basename(path))
         pred = imageio.imread(pred_path)
-        seg = _segment_impl(pred)
+        if sgn:
+            seg = _segment_impl(pred)
+        else:
+            seg = label(pred[0] > 0.5)
         v = napari.Viewer()
         v.add_image(image)
         v.add_image(pred)
         v.add_labels(seg)
+        v.title = os.path.basename(path)
         napari.run()
 
 
@@ -60,12 +83,15 @@ def check_segmentation(name):
 # /mnt/vast-nhr/home/pape41/u12086/Work/my_projects/flamingo-tools/scripts/training/checkpoints/cochlea_distance_unet_low-res-sgn  # noqa
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument("--model_path", "-m", required=True)
     parser.add_argument("--name", "-n", required=True)
+    parser.add_argument("--model_path", "-m")
+    parser.add_argument("--check", action="store_true")
+    parser.add_argument("--ihc", action="store_true")
     args = parser.parse_args()
 
-    # predict_blocks(args.model_path, args.name)
-    check_segmentation(args.name)
+    predict_blocks(args.model_path, args.name, sgn=not args.ihc)
+    if args.check:
+        check_segmentation(args.name, sgn=not args.ihc)
 
 
 if __name__ == "__main__":
